+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod audio;
 pub mod collision;
+pub mod colors;
 pub mod particles;
 
 pub mod prelude {